move vulkan creation functions into own cpp
