+++ Core/Graphics/Vulkan/VulkanCreation.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace Tk
+{
+namespace Core
+{
+namespace Graphics
+{
+
+void CreateSamplers();
+
+}
+}
+}
