+++ Core/Graphics/Vulkan/VulkanTypes.cpp
@@ -1,5 +1,4 @@
 #include "VulkanTypes.h"
-#include "Graphics/Vulkan/VulkanGPUMemAllocator.h"
 #include "Utility/Logging.h"
 
 #include <cstring>
@@ -11,109 +10,7 @@ namespace Core
 namespace Graphics
 {
 
-VulkanContextResources g_vulkanContextResources;
-
-void AllocGPUMemory(VkPhysicalDevice physicalDevice, VkDevice device, VkDeviceMemory* deviceMemory,
-    VkMemoryRequirements memRequirements, VkMemoryPropertyFlags memPropertyFlags)
-{
-    // Check for memory type support
-    VkPhysicalDeviceMemoryProperties memProperties;
-    vkGetPhysicalDeviceMemoryProperties(physicalDevice, &memProperties);
-
-    uint32 memTypeIndex = 0xffffffff;
-    for (uint32 uiMemType = 0; uiMemType < memProperties.memoryTypeCount; ++uiMemType)
-    {
-        if (((1 << uiMemType) & memRequirements.memoryTypeBits) &&
-            (memProperties.memoryTypes[uiMemType].propertyFlags & memPropertyFlags) == memPropertyFlags)
-        {
-            memTypeIndex = uiMemType;
-            break;
-        }
-    }
-    if (memTypeIndex == 0xffffffff)
-    {
-        Core::Utility::LogMsg("Platform", "Failed to find memory property flags!", Core::Utility::LogSeverity::eCritical);
-        TINKER_ASSERT(0);
-    }
-
-    VkMemoryAllocateInfo memAllocInfo = {};
-    memAllocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-    memAllocInfo.allocationSize = memRequirements.size;
-    memAllocInfo.memoryTypeIndex = memTypeIndex;
-    VkResult result = vkAllocateMemory(device, &memAllocInfo, nullptr, deviceMemory);
-    if (result != VK_SUCCESS)
-    {
-        Core::Utility::LogMsg("Platform", "Failed to allocate gpu memory!", Core::Utility::LogSeverity::eCritical);
-        TINKER_ASSERT(0);
-    }
-}
-
-void CreateBuffer(VkPhysicalDevice physicalDevice, VkDevice device, uint32 sizeInBytes,
-    VkBufferUsageFlags usageFlags, VkMemoryPropertyFlags propertyFlags,
-    VkBuffer* buffer, VkDeviceMemory* deviceMemory)
-{
-    VkBufferCreateInfo bufferCreateInfo = {};
-    bufferCreateInfo.sType = VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO;
-    bufferCreateInfo.size = sizeInBytes;
-    bufferCreateInfo.usage = usageFlags;
-    bufferCreateInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-
-    VkResult result = vkCreateBuffer(device, &bufferCreateInfo, nullptr, buffer);
-    if (result != VK_SUCCESS)
-    {
-        Core::Utility::LogMsg("Platform", "Failed to allocate Vulkan buffer!", Core::Utility::LogSeverity::eCritical);
-        TINKER_ASSERT(0);
-    }
-
-    VkMemoryRequirements memRequirements;
-    vkGetBufferMemoryRequirements(device, *buffer, &memRequirements);
-
-    AllocGPUMemory(physicalDevice, device, deviceMemory, memRequirements, propertyFlags);
-
-    vkBindBufferMemory(device, *buffer, *deviceMemory, 0);
-}
-
-void CreateImageView(VkDevice device, VkFormat format, VkImageAspectFlags aspectMask, VkImage image, VkImageView* imageView, uint32 arrayEles)
-{
-    VkImageViewCreateInfo imageViewCreateInfo = {};
-    imageViewCreateInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-    imageViewCreateInfo.image = image;
-    imageViewCreateInfo.viewType = arrayEles > 1 ? VK_IMAGE_VIEW_TYPE_2D_ARRAY: VK_IMAGE_VIEW_TYPE_2D;
-    imageViewCreateInfo.format = format;
-    imageViewCreateInfo.components.r = VK_COMPONENT_SWIZZLE_IDENTITY;
-    imageViewCreateInfo.components.g = VK_COMPONENT_SWIZZLE_IDENTITY;
-    imageViewCreateInfo.components.b = VK_COMPONENT_SWIZZLE_IDENTITY;
-    imageViewCreateInfo.components.a = VK_COMPONENT_SWIZZLE_IDENTITY;
-    imageViewCreateInfo.subresourceRange.aspectMask = aspectMask;
-    imageViewCreateInfo.subresourceRange.baseMipLevel = 0;
-    imageViewCreateInfo.subresourceRange.levelCount = 1;
-    imageViewCreateInfo.subresourceRange.baseArrayLayer = 0;
-    imageViewCreateInfo.subresourceRange.layerCount = arrayEles;
-
-    VkResult result = vkCreateImageView(device, &imageViewCreateInfo, nullptr, imageView);
-    if (result != VK_SUCCESS)
-    {
-        Core::Utility::LogMsg("Platform", "Failed to create Vulkan image view!", Core::Utility::LogSeverity::eCritical);
-        TINKER_ASSERT(0);
-    }
-}
-
-VkShaderModule CreateShaderModule(const char* shaderCode, uint32 numShaderCodeBytes, VkDevice device)
-{
-    VkShaderModuleCreateInfo shaderModuleCreateInfo = {};
-    shaderModuleCreateInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-    shaderModuleCreateInfo.codeSize = numShaderCodeBytes;
-    shaderModuleCreateInfo.pCode = reinterpret_cast<const uint32_t*>(shaderCode);
-
-    VkShaderModule shaderModule;
-    VkResult result = vkCreateShaderModule(device, &shaderModuleCreateInfo, nullptr, &shaderModule);
-    if (result != VK_SUCCESS)
-    {
-        Core::Utility::LogMsg("Platform", "Failed to create Vulkan shader module!", Core::Utility::LogSeverity::eCritical);
-        return VK_NULL_HANDLE;
-    }
-    return shaderModule;
-}
+VulkanContextResources g_vulkanContextResources = {};
 
 // NOTE: Must correspond the enums in PlatformGameAPI.h
 static VkPipelineColorBlendAttachmentState   VulkanBlendStates    [BlendState::eMax]     = {};
